+++ src/Instance/hooks/index.js
@@ -1,0 +1,12 @@
+export { default as useAuthoritiesByIdQuery } from './useAuthoritiesByIdQuery';
+export { default as useCirculationInstanceRequestsQuery } from './useCirculationInstanceRequestsQuery';
+export { default as useImportRecord } from './useImportRecord';
+export { default as useInstanceActions } from './useInstanceActions';
+export { default as useInstanceDetailsShortcuts } from './useInstanceDetailsShortcuts';
+export { default as useInstanceImportSupportedQuery } from './useInstanceImportSupportedQuery';
+export { default as useInstanceModalsContext } from './useInstanceModalsContext';
+export { default as useInstancePermissions } from './useInstancePermissions';
+export { default as useSharedInstancesQuery } from './useSharedInstancesQuery';
+export { default as useInstanceSharing } from './useInstanceSharing';
+export { default as useSetRecordForDeletion } from './useSetRecordForDeletion';
+export { default as useInstanceMutation } from './useInstanceMutation';