+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Item -> ViewItem: refactor component. Refs UIIN-3267.
 * *BREAKING* Use `/search/resources/jobs` endpoints to get Instances and Holdings UUIDs. Refs UIIN-3366.
 * Add source URI field to instance full-record display and (disabled) to edit form. Fixes UIIN-3287.
+* Instance -> ViewInstance: refactor component. Refs UIIN-3382.
 
 ## [13.0.4](https://github.com/folio-org/ui-inventory/tree/v13.0.4) (2025-04-18)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v13.0.3...v13.0.4)